+++ routes/core/detections/stream.js
@@ -1,7 +1,6 @@
 const router = require('express').Router()
 const { httpErrorHandler } = require('../../../utils/http-error-handler.js')
 const detectionsService = require('../../../services/detections')
-const annotationsService = require('../../../services/annotations')
 const classificationService = require('../../../services/classifications')
 const classifierService = require('../../../services/classifiers')
 const rolesService = require('../../../services/roles')
@@ -34,10 +33,6 @@ const auth0Service = require('../../../services/auth0/auth0-service')
  *         in: query
  *         required: true
  *         type: string
- *       - name: reviews
- *         description: Whether or not to include detection user reviews or not
- *         in: query
- *         type: boolean
  *       - name: classifications
  *         description: List of clasification identifiers
  *         in: query
@@ -58,17 +53,12 @@ const auth0Service = require('../../../services/auth0/auth0-service')
  *         default: 0
  *     responses:
  *       200:
- *         description: List of detections objects. **"reviews" attribute is included based on "reviews" query parameter**
  *         content:
  *           application/json:
  *             schema:
  *               type: array
  *               items:
- *                 oneOf:
- *                   - $ref: '#/components/schemas/Detection'
- *                   - $ref: '#/components/schemas/DetectionWithReviews'
- *                 discriminator:
- *                   propertyName: reviews
+ *                 $ref: '#/components/schemas/Detection'
  *       400:
  *         description: Invalid query parameters
  *       404:
@@ -85,7 +75,6 @@ router.get('/:id/detections', function (req, res) {
   params.convert('min_confidence').optional().toFloat()
   params.convert('limit').optional().toInt()
   params.convert('offset').optional().toInt()
-  params.convert('reviews').toBoolean().default(false)
 
   return params.validate()
     .then(async () => {
@@ -99,20 +88,9 @@ router.get('/:id/detections', function (req, res) {
       }
     })
     .then(async () => {
-      const { start, end, classifications, limit, offset, reviews } = convertedParams
+      const { start, end, classifications, limit, offset } = convertedParams
       const minConfidence = convertedParams.min_confidence
-      const proms = [detectionsService.query(start, end, streamId, classifications, minConfidence, limit, offset, user)]
-      if (reviews) {
-        proms.push(annotationsService.query({ start, end, streamId, classifications, user, isManual: false }, { limit, offset }))
-      }
-      const result = await Promise.all(proms)
-        .then(async (data) => {
-          let detections = data[0]
-          if (reviews) {
-            detections = await detectionsService.matchDetectionsWithReviews(data[0], data[1])
-          }
-          return detections
-        })
+      const result = await detectionsService.query(start, end, streamId, classifications, minConfidence, limit, offset, user)
       return res.json(result)
     })
     .catch(httpErrorHandler(req, res, 'Failed getting detections'))
@@ -194,8 +172,7 @@ router.post('/:id/detections', function (req, res) {
       // Get all the distinct classification values
       const classificationValues = [...new Set(validatedDetections.map(d => d.classification))]
       return classificationService.getIds(classificationValues)
-    })
-    .then(data => {
+    }).then(data => {
       classificationMapping = data
       const validatedDetections = params.transformedArray
       // Get all the distinct classifier uuids
@@ -204,21 +181,31 @@ router.post('/:id/detections', function (req, res) {
     }).then(classifierMapping => {
       const classifierIds = Object.values(classifierMapping)
       return Promise.all(classifierIds.map(id => classifierService.update(id, null, { last_executed_at: new Date() })))
-        .then(() => classifierMapping)
-    }).then(classifierMapping => {
+        .then(() => {
+          return Promise.all(classifierIds.map(id => classifierService.get(id, { joinRelations: true })))
+        })
+        .then((classifiers) => {
+          return Promise.resolve([classifiers, classifierMapping])
+        })
+    }).then(([classifiers, classifierMapping]) => {
       const validatedDetections = params.transformedArray
       const detections = validatedDetections.map(detection => {
         const classificationId = classificationMapping[detection.classification]
         const classifierId = classifierMapping[detection.classifier]
-        return {
-          streamId,
-          classificationId,
-          classifierId,
-          start: detection.start,
-          end: detection.end,
-          confidence: detection.confidence
+        const classifier = classifiers.find(c => c.id === classifierId)
+        const output = (classifier.outputs || []).find(i => i.classification_id === classificationId)
+        const threshold = output ? output.ignore_threshold : detectionsService.DEFAULT_IGNORE_THRESHOLD
+        if (detection.confidence > threshold) {
+          return {
+            streamId,
+            classificationId,
+            classifierId,
+            start: detection.start,
+            end: detection.end,
+            confidence: detection.confidence
+          }
         }
-      })
+      }).filter(i => i !== undefined)
       return detectionsService.create(detections)
     })
     .then(detections => res.sendStatus(201)) // TODO: not returning the ids of the created detections